Refactor :: device_name -> deviceName
