+++ src/main/java/com/keepgoing/keepserver/domain/device/payload/request/DeviceInfoRequest.java
@@ -12,7 +12,7 @@ public class DeviceInfoRequest {
     private Long id;
 
     /* 기기 이름 */
-    private String device_name;
+    private String deviceName;
 
     /* 기기 사진 */
     private String imgUrl;
