+++ src/main/java/com/keepgoing/keepserver/domain/device/payload/response/DeviceResponseDto.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 @Builder
 public record DeviceResponseDto(
         Long id,
-        String device_name,
+        String deviceName,
         String imgUrl,
         int status) {
 }
