+++ src/main/java/com/keepgoing/keepserver/domain/device/repository/DeviceRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface DeviceRepository extends JpaRepository<Device, Long> {
-    List<Device> findByDeviceUserNameContaining(String device_name, Sort id);
+    List<Device> findByDeviceNameContaining(String deviceName, Sort id);
 }
