+++ src/main/java/com/keepgoing/keepserver/domain/device/service/DeviceService.java
@@ -2,22 +2,17 @@ package com.keepgoing.keepserver.domain.device.service;
 
 import com.keepgoing.keepserver.domain.device.entity.Device;
 import com.keepgoing.keepserver.domain.device.payload.request.DeviceDto;
-import com.keepgoing.keepserver.domain.device.payload.request.DeviceRequest;
 import com.keepgoing.keepserver.domain.device.payload.response.DeviceResponseDto;
 import com.keepgoing.keepserver.global.dto.response.BaseResponse;
-import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
+import org.springframework.security.core.Authentication;
 
 public interface DeviceService {
-
-    ResponseEntity<BaseResponse> deviceCreate(DeviceRequest request, MultipartFile multipartFile , Authentication authentication) throws IOException;
 
     ResponseEntity<BaseResponse> myDevices(Authentication authentication);
 
     ResponseEntity<BaseResponse> deviceRead(Long id);
+    ResponseEntity<BaseResponse> deleteDevice(Long id, Authentication authentication);
 
     ResponseEntity<BaseResponse> findAll();
 
@@ -25,7 +20,7 @@ public interface DeviceService {
 
     default Device dtoToEntity(DeviceDto dto){
         return Device.builder()
-                .device_name(dto.getDevice_name())
+                .deviceName(dto.getDeviceName())
                 .status(dto.getStatus())
                 .imgUrl(dto.getImgUrl())
                 .build();
@@ -34,7 +29,7 @@ public interface DeviceService {
     default DeviceResponseDto entityToDto(Device entity){
         return DeviceResponseDto.builder()
                 .id(entity.getId())
-                .device_name((entity.getDevice_name()))
+                .deviceName((entity.getDeviceName()))
                 .status(entity.getStatus())
                 .build();
     }
