+++ src/main/java/com/keepgoing/keepserver/domain/user/payload/request/SignupRequest.java
@@ -3,8 +3,6 @@ package com.keepgoing.keepserver.domain.user.payload.request;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.List;
-
 @Getter
 @Setter
 public class SignupRequest {
